+++ src/julia_to_gap.jl
@@ -1,31 +1,53 @@
 ## Converters
 """
-    julia_to_gap(input, recursion_dict = IdDict(); recursive::Bool = false)
+    GapObj(input, recursion_dict::GapCacheDict = nothing; recursive::Bool = false)
+    GapObj(input, recursive::Bool = false)
 
-Convert a julia object `input` to an appropriate GAP object.
-If `recursive` is set to `true`, recursive conversions on
-arrays, tuples, and dictionaries is performed.
+One can use the type [`GapObj`](@ref) as a constructor,
+in order to convert the julia object `input` to an appropriate GAP object.
+
+If `recursive` is set to `true`, recursive conversion of nested Julia objects
+(arrays, tuples, and dictionaries) is performed.
+
 
 The input `recursion_dict` should never be set by the user, it is meant to keep egality
 of input data, by converting equal data to identical objects in GAP.
 
 # Examples
 ```jldoctest
-julia> GAP.julia_to_gap(1//3)
+julia> GapObj(1//3)
 GAP: 1/3
 
-julia> GAP.julia_to_gap("abc")
+julia> GapObj("abc")
 GAP: "abc"
 
-julia> GAP.julia_to_gap([ [1, 2], [3, 4]])
-GAP: [ <Julia: [1, 2]>, <Julia: [3, 4]> ]
-
-julia> GAP.julia_to_gap([ [1, 2], [3, 4]], recursive = true)
+julia> GapObj([1 2; 3 4])
 GAP: [ [ 1, 2 ], [ 3, 4 ] ]
 
+julia> GapObj([[1, 2], [3, 4]])
+GAP: [ <Julia: [1, 2]>, <Julia: [3, 4]> ]
+
+julia> GapObj([[1, 2], [3, 4]], true)
+GAP: [ [ 1, 2 ], [ 3, 4 ] ]
+
+julia> GapObj([[1, 2], [3, 4]], recursive = true)
+GAP: [ [ 1, 2 ], [ 3, 4 ] ]
 ```
 
-The following `julia_to_gap` conversions are supported by GAP.jl.
+Note that this conversion is *not* restricted to outputs that actually are
+of type `GapObj`,
+also GAP integers, finite field elements, and booleans can be created
+by the constructor `GapObj`.
+
+```jldoctest
+julia> res = GapObj(42);  res isa GapObj
+false
+
+julia> res isa GAP.Obj
+true
+```
+
+The following `GapObj` conversions are supported by GAP.jl.
 (Other Julia packages may provide conversions for more Julia objects.)
 
 | Julia type                           | GAP filter   |
@@ -40,50 +62,66 @@ The following `julia_to_gap` conversions are supported by GAP.jl.
 | `Char`                               | `IsChar`     |
 | `Vector{T}`                          | `IsList`     |
 | `Vector{Bool}`, `BitVector`          | `IsBList`    |
+| `Set{T}`                             | `IsList`     |
 | `Tuple{T}`                           | `IsList`     |
 | `Matrix{T}`                          | `IsList`     |
 | `Dict{String, T}`, `Dict{Symbol, T}` | `IsRecord`   |
 | `UnitRange{T}`, `StepRange{T, S}`    | `IsRange`    |
 | `Function`                           | `IsFunction` |
 """
-julia_to_gap(x::FFE) = x    # Default for actual GAP objects is to do nothing
-julia_to_gap(x::Bool) = x   # Default for actual GAP objects is to do nothing
+GapObj(x, cache::GapCacheDict = nothing; recursive::Bool = false) = GapObj_internal(x, cache, Val(recursive))
 
-## Integers: general case first deal with things that fit into immediate
-## integers, then falls back to converting to BigInt and calling into the GAP
-## kernel API.
-## TODO: we could provide more efficient conversion for UInt64, Int128, UInt128
-## which avoids the conversion to BigInt, if we wanted to.
-function julia_to_gap(x::Integer)
-    # if it fits into a GAP immediate integer, convert x to Int64
-    x in -1<<60:(1<<60-1) && return Int64(x)
-    # for the general case, fall back to BigInt
-    return julia_to_gap(BigInt(x))
-end
+# The calls to `GAP.@install` install methods for `GAP.GapObj_internal`
+# so we must make sure it is declared before
+function GapObj_internal end
+
+
+# The idea behind `_needs_tracking_julia_to_gap` is to avoid the creation
+# of a dictionary for tracking subobjects if their type implies that
+# no such tracking is needed/wanted.
+#
+# We need anyhow a `_needs_tracking_julia_to_gap` method for `Any`,
+# for example in order to convert a Julia object of type `Dict{Symbol, Any}`
+# to a GAP record.
+# Thus we have a default method returning `true`.
+#
+# Methods for those types that want `false` have to be installed explicitly.
+# The `GapObj` methods arising from `GAP.@install` don't handle recursion
+# and the macro automatically installs such a `_needs_tracking_julia_to_gap`
+# method.
+_needs_tracking_julia_to_gap(::Any) = true
+
+
+GAP.@install GapObj(x::FFE) = x    # Default for actual GAP objects is to do nothing
+GAP.@install GapObj(x::Bool) = x    # Default for actual GAP objects is to do nothing
+
+## Integers:
+## We do not want to track conversion for any concrete integer types.
+@install GapObj(x::Integer) = x in -1<<60:(1<<60-1) ? Int64(x) : GapObj(BigInt(x))
 
 ## Small integers types always fit into GAP immediate integers, and thus are
 ## represented by Int64 on the Julia side.
-julia_to_gap(x::Int64) = x
-julia_to_gap(x::Int32) = Int64(x)
-julia_to_gap(x::Int16) = Int64(x)
-julia_to_gap(x::Int8) = Int64(x)
-julia_to_gap(x::UInt32) = Int64(x)
-julia_to_gap(x::UInt16) = Int64(x)
-julia_to_gap(x::UInt8) = Int64(x)
+GAP.@install GapObj(x::Int64) = x
+GAP.@install GapObj(x::Int32) = Int64(x)
+GAP.@install GapObj(x::Int16) = Int64(x)
+GAP.@install GapObj(x::Int8) = Int64(x)
+GAP.@install GapObj(x::UInt32) = Int64(x)
+GAP.@install GapObj(x::UInt16) = Int64(x)
+GAP.@install GapObj(x::UInt8) = Int64(x)
 
-function julia_to_gap(x::UInt)
+GAP.@install function GapObj(x::UInt)
     x < (1<<60) && return Int64(x)
     return ccall((:ObjInt_UInt, libgap), GapObj, (UInt64, ), x)
 end
 
 ## BigInts are converted via a ccall
-function julia_to_gap(x::BigInt)
+GAP.@install function GapObj(x::BigInt)
     x in -1<<60:(1<<60-1) && return Int64(x)
     return GC.@preserve x ccall((:MakeObjInt, libgap), GapObj, (Ptr{UInt64}, Cint), x.d, x.size)
 end
 
 ## Rationals
-function julia_to_gap(x::Rational{T}) where {T<:Integer}
+function GapObj_internal(x::Rational{T}, cache::GapCacheDict, ::Val{recursive}) where {T<:Integer, recursive}
     denom_julia = denominator(x)
     numer_julia = numerator(x)
     if denom_julia == 0
@@ -93,148 +131,213 @@ function julia_to_gap(x::Rational{T}) where {T<:Integer}
             return -Globals.infinity
         end
     end
-    numer = julia_to_gap(numer_julia)
-    denom = julia_to_gap(denom_julia)
+    numer = GapObj_internal(numer_julia, cache, Val(recursive))
+    denom = GapObj_internal(denom_julia, cache, Val(recursive))
     return Wrappers.QUO(numer, denom)
 end
 
 ## Floats
-julia_to_gap(x::Float64) = NEW_MACFLOAT(x)
-julia_to_gap(x::Float32) = NEW_MACFLOAT(Float64(x))
-julia_to_gap(x::Float16) = NEW_MACFLOAT(Float64(x))
+GAP.@install GapObj(x::Float64) = NEW_MACFLOAT(x)
+GAP.@install GapObj(x::Float32) = NEW_MACFLOAT(Float64(x))
+GAP.@install GapObj(x::Float16) = NEW_MACFLOAT(Float64(x))
 
 ## Chars
-julia_to_gap(x::Char) = CharWithValue(Cuchar(x))
+GAP.@install GapObj(x::Char) = CharWithValue(Cuchar(x))
 
 ## Strings and symbols
-julia_to_gap(x::AbstractString) = MakeString(string(x))
-julia_to_gap(x::Symbol) = MakeString(string(x))
+GAP.@install GapObj(x::AbstractString) = MakeString(string(x))
+GAP.@install GapObj(x::Symbol) = MakeString(string(x))
 
-## Generic caller for optional arguments
-julia_to_gap(obj::Any; recursive::Bool) = julia_to_gap(obj)
-julia_to_gap(obj::Any, recursion_dict::IdDict{Any,Any}; recursive::Bool = true) = julia_to_gap(obj)
+# helper functions for recursion
+function recursion_info(::Type{T}, obj, recursive::Bool, recursion_dict::GapCacheDict) where {T}
+    recursive && recursion_dict !== nothing && haskey(recursion_dict, obj) && return true, false, recursion_dict
+    rec = recursive && _needs_tracking_julia_to_gap(T)
+    if rec && recursion_dict === nothing
+        rec_dict = RecDict()
+    else
+        rec_dict = recursion_dict
+    end
+
+    return false, rec, rec_dict
+end
+
+function handle_recursion(obj, ret_val, rec::Bool, rec_dict::GapCacheDict)
+    if rec_dict !== nothing
+        rec_dict[obj] = ret_val
+    end
+    return rec ? rec_dict : nothing
+end
 
 ## Arrays (including BitVector)
-function julia_to_gap(
+function GapObj_internal(
     obj::AbstractVector{T},
-    recursion_dict::IdDict{Any,Any} = IdDict();
-    recursive::Bool = false,
-) where {T}
+    recursion_dict::GapCacheDict,
+    ::Val{recursive},
+) where {T, recursive}
+
+    ret, rec, rec_dict = recursion_info(T, obj, recursive, recursion_dict)
+    ret && return rec_dict[obj]
 
     len = length(obj)
     ret_val = NewPlist(len)
-    if recursive
-        recursion_dict[obj] = ret_val
-    end
+
+    recursion_dict = handle_recursion(obj, ret_val, rec, rec_dict)
+
+    # Set the subobjects.
     for i = 1:len
         x = obj[i]
         if x === nothing
             continue
         end
         if recursive
-            x = get!(recursion_dict, x) do
-                julia_to_gap(x, recursion_dict; recursive)
-            end
+            # Convert the subobjects.
+            # Do not care about findíng them in the dictionary
+            # or adding them to the dictionary,
+            # since their conversion method decides about that.
+            res = GAP.GapObj_internal(x, recursion_dict, Val(true))
+        else
+            res = x
         end
-        ret_val[i] = x
+        ret_val[i] = res
     end
+
+    return ret_val
+end
+
+## Sets
+function GapObj_internal(
+    obj::Set{T},
+    recursion_dict::GapCacheDict,
+    ::Val{recursive},
+) where {T, recursive}
+
+    ret, rec, rec_dict = recursion_info(T, obj, recursive, recursion_dict)
+    ret && return rec_dict[obj]
+
+    ret_val = NewPlist(length(obj))
+
+    recursion_dict = handle_recursion(obj, ret_val, rec, rec_dict)
+
+    for x in obj
+        if recursive
+            res = GapObj_internal(x, recursion_dict, Val(true))
+        else
+            res = x
+        end
+        Wrappers.Add(ret_val, res)
+    end
+    Wrappers.Sort(ret_val)
+    @assert Wrappers.IsSet(ret_val)
+
     return ret_val
 end
 
 ## Convert two dimensional arrays
-function julia_to_gap(
+function GapObj_internal(
     obj::Matrix{T},
-    recursion_dict::IdDict{Any,Any} = IdDict();
-    recursive::Bool = false,
-) where {T}
-    (rows, cols) = size(obj)
-    if haskey(recursion_dict, obj)
-        return recursion_dict[obj]
-    end
+    recursion_dict::GapCacheDict,
+    ::Val{recursive},
+) where {T, recursive}
+
+    ret, rec, rec_dict = recursion_info(T, obj, recursive, recursion_dict)
+    ret && return rec_dict[obj]
+
+    rows = size(obj, 1)
     ret_val = NewPlist(rows)
-    if recursive
-        recursion_dict[obj] = ret_val
-    end
+
+    recursion_dict = handle_recursion(obj, ret_val, rec, rec_dict)
+
     for i = 1:rows
-        ret_val[i] = julia_to_gap(obj[i, :], recursion_dict; recursive)
+        # We need not distinguish between recursive or not
+        # because we are just now creating the "row objects" in Julia.
+        ret_val[i] = GapObj_internal(obj[i, :], recursion_dict, Val(true))
     end
     return ret_val
 end
 
 ## Tuples
-function julia_to_gap(
+function GapObj_internal(
     obj::Tuple,
-    recursion_dict::IdDict{Any,Any} = IdDict();
-    recursive::Bool = false,
-)
+    recursion_dict::GapCacheDict,
+    ::Val{recursive},
+) where recursive
     array = collect(Any, obj)
-    return julia_to_gap(array, recursion_dict; recursive)
+    return GapObj_internal(array, recursion_dict, Val(recursive))
 end
 
 ## Ranges
-function julia_to_gap(r::AbstractRange{<:Integer}, recursive::Bool = false)
+GAP.@install function GapObj(r::AbstractRange{<:Integer})
     res = NewRange(length(r), first(r), step(r))
     Wrappers.IsRange(res) || throw(ConversionError(r, GapObj))
     return res
 end
 
-function julia_to_gap(
-    r::AbstractRange{<:Integer},
-    recursion_dict::IdDict{Any,Any};
-    recursive::Bool = false)
-
-    return julia_to_gap(r)
-end
-
 ## Dictionaries
-function julia_to_gap(
-    obj::Dict{T,S},
-    recursion_dict::IdDict{Any,Any} = IdDict();
-    recursive::Bool = false,
-) where {S} where {T<:Union{Symbol,AbstractString}}
+function GapObj_internal(
+    obj::Dict{S,T},
+    recursion_dict::GapCacheDict,
+    ::Val{recursive},
+) where {T, S<:Union{Symbol,AbstractString}, recursive}
 
-    record = NewPrecord(0)
-    if recursive
-        recursion_dict[obj] = record
-    end
+    ret, rec, rec_dict = recursion_info(T, obj, recursive, recursion_dict)
+    ret && return rec_dict[obj]
+
+    ret_val = NewPrecord(0)
+
+    recursion_dict = handle_recursion(obj, ret_val, rec, rec_dict)
+
     for (x, y) in obj
         x = Wrappers.RNamObj(MakeString(string(x)))
         if recursive
-            y = get!(recursion_dict, y) do
-                julia_to_gap(y, recursion_dict; recursive)
-            end
+            res = GapObj_internal(y, recursion_dict, Val(true))
+        else
+            res = y
         end
-        Wrappers.ASS_REC(record, x, y)
+        Wrappers.ASS_REC(ret_val, x, res)
     end
 
-    return record
+    return ret_val
 end
 
 ## GAP objects:
 ## We have to do something only if recursive conversion is required,
 ## and if `obj` contains Julia subobjects;
 ## in this case, `obj` is a GAP list or record.
-## An example of such an `obj` is `GAP.julia_to_gap([[1]])`.
-function julia_to_gap(
+## An example of such an `obj` is `GapObj([[1]])`.
+function GapObj_internal(
     obj::GapObj,
-    recursion_dict::IdDict{Any,Any} = IdDict();
-    recursive::Bool = false,
-)
+    recursion_dict::GapCacheDict,
+    ::Val{recursive},
+) where {recursive}
+
+    recursive && recursion_dict !== nothing && haskey(recursion_dict, obj) && return recursion_dict[obj]
+
     if ! recursive
         ret_val = obj
     elseif Wrappers.IsList(obj)
         len = length(obj)
         ret_val = NewPlist(len)
+        if recursion_dict === nothing
+            # We have no type information that allows us to avoid the dictionary.
+            recursion_dict = RecDict()
+        end
         recursion_dict[obj] = ret_val
         for i = 1:len
-             ret_val[i] = julia_to_gap(obj[i], recursion_dict; recursive)
+            x = obj[i]
+            ret_val[i] = GapObj_internal(x, recursion_dict::RecDict, Val(recursive))
         end
     elseif Wrappers.IsRecord(obj)
         ret_val = NewPrecord(0)
+        if recursion_dict === nothing
+            # We have no type information that allows us to avoid the dictionary.
+            recursion_dict = RecDict()
+        end
         recursion_dict[obj] = ret_val
         for xx in Wrappers.RecNames(obj)::GapObj
             x = Wrappers.RNamObj(xx)
-            Wrappers.ASS_REC(ret_val, x, julia_to_gap(Wrappers.ELM_REC(obj, x), recursion_dict; recursive = true))
+            y = Wrappers.ELM_REC(obj, x)
+            res = GapObj_internal(y, recursion_dict::RecDict, Val(recursive))
+            Wrappers.ASS_REC(ret_val, x, res)
         end
     else
         ret_val = obj
@@ -243,4 +346,4 @@ function julia_to_gap(
     return ret_val
 end
 
-julia_to_gap(func::Function) = WrapJuliaFunc(func)
+GAP.@install GapObj(func::Function) = WrapJuliaFunc(func)