+++ test/convenience.jl
@@ -68,7 +68,7 @@ end
 
     # Julia object in GAP list
     l = [ [ 1 2 ], [ 3 4 ] ]
-    gaplist = GAP.julia_to_gap( l )  # list of Julia arrays
+    gaplist = GapObj( l )  # list of Julia arrays
     @test l[1] in gaplist
     @test !([ 5 6 ] in gaplist)
 end