+++ CHANGES.md
@@ -1,5 +1,13 @@
 # Changes in GAP.jl
 
+- Rewrite `julia_to_gap`, in order to
+  - make the installation of new conversion methods from Julia to GAP
+    simpler and safer and
+  - restrict the necessity to create dictionaries to situations where
+    recursive conversions make sense.
+  For that, the function `GAP.GapObj_internal`, the macro `GAP.@install`,
+  and the type `GapCacheDict` were introduced.
+
 ## Version 0.11.4 (released 2024-09-19)
 
 - Support AbstractAlgebra 0.43